Refactored estimator to be a class to simplify main function

+++ functions/Estimation/GCTA_wrapper.py
@@ -8,69 +8,53 @@ Created on Mon Nov 21 07:35:15 2022
 import os
 import subprocess
 import pandas as pd 
+from functions.Data_input.load_data import load_everything
 # os.chdir("/home/christian/Research/Stat_gen/tools/Basu_herit")
 
 #%%
-
-def GCTA(grm, pheno_file, cov_file, PC_file, covars, nnpc, mp) : 
-    """
-    A wrapper for the GCTA GREML heritability estimator.
-
-    Parameters
-    ----------
-    grm : string
-        file path prefix to GRM files.
-    pheno_file : string
-        file path to phenotype file.
-    cov_file : string
-        file path to covariates file.
-    PC_file : string
-        file path to covariates file.
-    covars : list of strings
-        list of covaraites to control for as fixed effects.
-    nnpc : int
-        number of pc's to control for as fixed effects.
-    mp : string
-        name of phenotype to estimate on.
-
-    Returns
-    -------
-    pandas dataframe.
-        Dataframe of estimated heritability and standard error
-
-    """
-    # Load and select covariates
-    eigs = pd.read_csv(PC_file, sep = " ", header = None)
-    cols = eigs.columns.tolist()[2:]
-    eigs.columns = ["fid", "iid"] + [i -1  for i in cols]
-    eigs = eigs.iloc[:,:nnpc+2]
-    cov = pd.read_csv(cov_file, sep = " ", header= 0)
-    cov.columns = [col.lower() for col in cov.columns.tolist()]
-    cov = cov[["fid", "iid"] + covars]
-    cov = cov.merge(eigs, on = ["fid", "iid"])
-    # Decide which are qcovars and which are numeric covars
-    discrete = [len(cov[col].unique()) < 50 for col in cov]
-    # Include FID IID
-    discrete[0:2] = [True, True]
-    cont = [not v for v in discrete]
-    cont[0:2] = [True, True]
+def GCTA(df, covars, nnpc, mp, GRM, std = False, Method = "AdjHE", RV = None, silent=False):
+    # write the phenotype file
+    df[["fid", "iid", mp]].to_csv("temp_pheno.txt", sep = " ", header = False, index= False, na_rep = "NA")
     
-    # Svae temp files
-    cov.iloc[:,discrete].to_csv("temp_Discrete.txt", sep = " ", header = False, index= False)
-    cov.iloc[:,cont].to_csv("temp_Cont.txt", sep = " ", header= False, index= False)
+    # Select the remaining variables of interest
+    pcs =  ["pc_" + str(s) for s in range(1, nnpc)]
+    df = df[["fid", "iid"] + covars + pcs]
+
+
+    # Decide which are qcovars and which are numeric covars
+    discrete = [(len(df[col].unique()) < 50) and (len(df[col].unique()) > 1) for col in df]
+    # Include FID IID
+    cont = [not v for v in discrete]
+    discrete[0:2] = [True, True]
+
+    
+    # Svae temp files if there were any covariates in either category
+    if sum(discrete) > 2:
+        df.iloc[:,discrete].to_csv("temp_Discrete.txt", sep = " ", header = False, index= False, na_rep = "NA")
+    if sum(cont) > 2:
+        df.iloc[:,cont].to_csv("temp_Cont.txt", sep = " ", header= False, index= False, na_rep = "NA")
+        
+    # Format string for controlling variables
+    covars = " "
+    if os.path.exists("temp_Cont.txt") :
+        covars += " --qcovar temp_Cont.txt "
+    if os.path.exists("temp_Discrete.txt") : 
+        covars += " --covar temp_Discrete.txt "
+    
     # Find GCTA
     gcta = "whereis gcta64"
     gcta, __ = subprocess.Popen(
         gcta.split(), stdout=subprocess.PIPE).communicate()
     gcta= gcta.split()[1].decode("utf-8")
+
+    
     # run gcta
-    out = "test"
-    bashcommand = gcta + f" --grm {grm} --pheno {pheno_file} --mpheno 1 --reml --out {out} --qcovar temp_Cont.txt --covar temp_Discrete.txt"
+    bashcommand = gcta + f" --grm {GRM} --pheno temp_pheno.txt --mpheno 1 --reml --out temp" + covars
     process = subprocess.Popen(bashcommand.split(), stdout=subprocess.PIPE)
     __output, __error = process.communicate()
 
-    #
-    df = pd.read_table(out + ".hsq", sep="\t").query( "Source == 'V(G)/Vp'").reset_index()
+    # parse output for estimate
+    df = pd.read_table("temp" + ".hsq", sep="\t").query( "Source == 'V(G)/Vp'").reset_index()
     
     result = {"h2" : df.Variance[0],
               "SE" : df.SE[0],
@@ -81,12 +65,15 @@ def GCTA(grm, pheno_file, cov_file, PC_file, covars, nnpc, mp) :
               "Memory Usage" : 0}
     
     # tidy up by removing temporary files
-    os.remove("temp_Discrete.txt")
-    os.remove("temp_Cont.txt")
+    if os.path.exists("temp_Discrete.txt") : 
+        os.remove("temp_Discrete.txt")
+    if os.path.exists("temp_Discrete.txt") : 
+        os.remove("temp_Cont.txt")
     
     # Return the fit results
-    return(pd.DataFrame(result, index = [0]))
+    return pd.DataFrame(result, index = [0])
 
+
